backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.7.4
email-validator==2.2.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
prophet==1.1.5
statsmodels==0.14.2
pdfplumber==0.11.0
openpyxl==3.1.5
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.6.0
psycopg2-binary==2.9.9
razorpay==1.4.2
